refactor code with zustand
