+++ components/Loading.js
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import {Text, View} from 'react-native';
 import LottieView from 'lottie-react-native';
+
 import COLORS from '../color';
 
 export default function Loading() {
