+++ components/collection/FavButton.js
@@ -1,16 +1,14 @@
 import * as React from 'react';
-import {useRef, useEffect} from 'react';
 import {StyleSheet, TouchableOpacity} from 'react-native';
 import LottieView from 'lottie-react-native';
-import {useFocusEffect} from '@react-navigation/native';
 import {
   widthPercentageToDP as wp,
   heightPercentageToDP as hp,
 } from 'react-native-responsive-screen';
 
 function usePrevious(value) {
-  const ref = useRef();
-  useEffect(() => {
+  const ref = React.useRef();
+  React.useEffect(() => {
     ref.current = value;
   });
   return ref.current;
@@ -30,7 +28,7 @@ export default function FavButton(props) {
       <TouchableOpacity style={styles.favTouchArea} onPress={onPress} key="1">
         <LottieView
           style={styles.fav}
-          source={require('../src/2415-twitter-heart.json')}
+          source={require('../../src/2415-twitter-heart.json')}
           autoPlay={true}
           loop={false}
           progress={0.3}
@@ -42,7 +40,7 @@ export default function FavButton(props) {
       <TouchableOpacity style={styles.favTouchArea} onPress={onPress} key="2">
         <LottieView
           style={styles.fav}
-          source={require('../src/2415-twitter-heart.json')}
+          source={require('../../src/2415-twitter-heart.json')}
           progress={isFavorite ? 1 : 0}
         />
       </TouchableOpacity>
