+++ components/collection/Index.js
@@ -1,26 +1,28 @@
 import * as React from 'react';
-import {useState} from 'react';
 import {ScrollView} from 'react-native-gesture-handler';
 import {StyleSheet, Text, TouchableOpacity, Image, View} from 'react-native';
 import {Container, Header, Left, Body, Right, Title} from 'native-base';
 import {useFocusEffect} from '@react-navigation/native';
-import utils from '../utils.js';
-import Loading from './Loading.js';
-import FavButton from './FavButton.js';
-import COLORS from '../color';
 import {
   widthPercentageToDP as wp,
   heightPercentageToDP as hp,
 } from 'react-native-responsive-screen';
 import LinearGradient from 'react-native-linear-gradient';
 import Icon from 'react-native-vector-icons/FontAwesome5';
-import useZ from '../zustand';
+
+import utils from '../../utils.js';
+import COLORS from '../../color';
+
+import Loading from '../Loading.js';
+import FavButton from './FavButton.js';
+
+import {useZ} from '../../zustand';
 
 export default function Collection({route}) {
   const popUpWaifu = useZ(z => z.popUpWaifu);
-  const [loading, setLoading] = useState(true);
-  const [collection, setCollection] = useState([]);
-  const [selectedWaifuIdx, setSelectedWaifuIdx] = useState(0);
+  const [loading, setLoading] = React.useState(true);
+  const [collection, setCollection] = React.useState([]);
+  const [selectedWaifuIdx, setSelectedWaifuIdx] = React.useState(0);
 
   useFocusEffect(
     React.useCallback(() => {
